Refactor test lib into tests/common
