+++ models/demos/blackhole/resnet50/tests/upstream_pipeline/test_perf_e2e_resnet50_batch32.py
@@ -5,7 +5,7 @@
 import pytest
 
 from models.common.utility_functions import run_for_blackhole
-from models.demos.ttnn_resnet.test_lib.perf_e2e_resnet50 import run_perf_resnet
+from models.demos.ttnn_resnet.tests.common.perf_e2e_resnet50 import run_perf_resnet
 
 
 @run_for_blackhole()
