+++ models/demos/classification_eval/classification_eval.py
@@ -250,7 +250,7 @@ def test_vit_image_classification_eval(
 def run_resnet50_image_classification_eval(
     device, model_type, batch_size_per_device, act_dtype, weight_dtype, model_location_generator, resolution=224
 ):
-    from models.demos.ttnn_resnet.test_lib.resnet50_test_infra import create_test_infra
+    from models.demos.ttnn_resnet.tests.common.resnet50_test_infra import create_test_infra
     from models.demos.utils.common_demo_utils import get_batch
     from models.tt_cnn.tt.pipeline import PipelineConfig, create_pipeline_from_config
