+++ models/demos/grayskull/resnet50/tests/test_perf_device_resnet50.py
@@ -5,7 +5,7 @@
 import pytest
 
 from models.common.utility_functions import run_for_grayskull
-from models.demos.ttnn_resnet.test_lib.perf_device_resnet50 import run_perf_device
+from models.demos.ttnn_resnet.tests.common.perf_device_resnet50 import run_perf_device
 
 
 @run_for_grayskull()
