+++ models/demos/ttnn_resnet/tests/common/perf_e2e_resnet50.py
