+++ models/demos/wormhole/resnet50/tests/test_resnet50_functional.py
@@ -6,7 +6,7 @@ import pytest
 
 import ttnn
 from models.common.utility_functions import is_blackhole
-from models.demos.ttnn_resnet.test_lib.resnet50_test_infra import create_test_infra
+from models.demos.ttnn_resnet.tests.common.resnet50_test_infra import create_test_infra
 
 
 def run_resnet_50(
