+++ models/demos/wormhole/resnet50/tests/test_resnet50_performant.py
@@ -6,7 +6,7 @@ import pytest
 
 import ttnn
 from models.common.utility_functions import run_for_wormhole_b0
-from models.demos.ttnn_resnet.test_lib.resnet50_performant import (
+from models.demos.ttnn_resnet.tests.common.resnet50_performant import (
     run_resnet50_2cqs_inference,
     run_resnet50_inference,
     run_resnet50_trace_2cqs_inference,
