+++ models/demos/wormhole/resnet50/tests/test_resnet50_performant_imagenet.py
@@ -10,8 +10,8 @@ from transformers import AutoImageProcessor
 
 import ttnn
 from models.common.utility_functions import profiler, run_for_wormhole_b0
-from models.demos.ttnn_resnet.test_lib.demo_utils import get_batch, get_data_loader
-from models.demos.ttnn_resnet.test_lib.resnet50_test_infra import create_test_infra
+from models.demos.ttnn_resnet.tests.common.demo_utils import get_batch, get_data_loader
+from models.demos.ttnn_resnet.tests.common.resnet50_test_infra import create_test_infra
 from models.tt_cnn.tt.pipeline import PipelineConfig, create_pipeline_from_config
 
 NUM_VALIDATION_IMAGES_IMAGENET = 49920
