+++ tests/nightly/single_card/resnet50/test_ttnn_functional_resnet50_large.py
@@ -21,7 +21,7 @@ from models.common.utility_functions import (
     is_wormhole_b0,
 )
 
-from models.demos.ttnn_resnet.test_lib.resnet50_test_infra import load_resnet50_model
+from models.demos.ttnn_resnet.tests.common.resnet50_test_infra import load_resnet50_model
 from models.demos.ttnn_resnet.tt.ttnn_functional_resnet50_large import resnet50
